pyproject.toml
[project]
name = "tele-home-supervisor"
version = "0.1.0"
description = "A Telegram bot for home server supervision."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=22.5",
    "psutil>=7.1.3",
    "docker>=7.1.0",
    "requests>=2.32.5",
    "urllib3>=2.6.2",
    "qbittorrent-api>=2025.11.1",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
